refactor(vllm): organize custom encoder modules
