+++ components/src/dynamo/vllm/handlers.py
@@ -90,9 +90,10 @@ from .cache_info import get_configured_kv_event_block_size
 from .constants import DisaggregationMode, EmbeddingTransferMode
 from .engine_monitor import VllmEngineMonitor
 from .lora_state import LoRAState
-from .multimodal_utils.async_vision_encoder import AsyncVisionEncoder
-from .multimodal_utils.custom_encoder_adapter import (
+from .multimodal_utils.custom_encoder import (
+    AsyncVisionEncoder,
     CustomEncoderAdapter,
+    VisionEncoderBackend,
     create_custom_encoder_adapter,
 )
 from .multimodal_utils.prefill_worker_utils import MultiModalEmbeddingLoader
@@ -102,7 +103,6 @@ from .multimodal_utils.request_processor import (
     MissingMultimodalHandoffError,
     VllmMultimodalRequestProcessor,
 )
-from .multimodal_utils.vision_encoder_backend import VisionEncoderBackend
 
 configure_dynamo_logging()
 logger = logging.getLogger(__name__)
