+++ components/src/dynamo/vllm/multimodal_utils/custom_encoder/async_encoder.py
@@ -34,13 +34,13 @@ from typing import Generic, List
 
 import torch
 
-from dynamo.vllm.multimodal_utils.threaded_micro_batcher import ThreadedMicroBatcher
-from dynamo.vllm.multimodal_utils.vision_encoder_backend import (
+from dynamo.vllm.multimodal_utils.custom_encoder.backend import (
     ItemT,
     Preprocessed,
     RawT,
     VisionEncoderBackend,
 )
+from dynamo.vllm.multimodal_utils.custom_encoder.batcher import ThreadedMicroBatcher
 
 
 class AsyncVisionEncoder(Generic[RawT, ItemT]):
