+++ components/src/dynamo/vllm/multimodal_utils/custom_encoder/backend/__init__.py
@@ -1,0 +1,18 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Author-facing custom encoder backend contract."""
+
+from dynamo.vllm.multimodal_utils.custom_encoder.backend.base import (
+    ItemT,
+    Preprocessed,
+    RawT,
+    VisionEncoderBackend,
+)
+
+__all__ = [
+    "ItemT",
+    "Preprocessed",
+    "RawT",
+    "VisionEncoderBackend",
+]
