+++ components/src/dynamo/vllm/tests/test_vllm_custom_encoder_handler.py
@@ -8,10 +8,10 @@ import pytest
 import torch
 
 from dynamo.vllm.handlers import DecodeWorkerHandler
-from dynamo.vllm.multimodal_utils.custom_encoder_adapter import (
+from dynamo.vllm.multimodal_utils.custom_encoder import (
+    VisionEncoderBackend,
     create_custom_encoder_adapter,
 )
-from dynamo.vllm.multimodal_utils.vision_encoder_backend import VisionEncoderBackend
 
 pytestmark = [
     pytest.mark.unit,
